+++ include/H2DE/H2DE.h
@@ -1,5 +1,5 @@
 /**
- * Version: 1.3.11
+ * Version: 1.3.12
  */
 
 #include "H2DE_engine.h"