+++ src/beatmap.cpp
@@ -4,14 +4,17 @@
 Beatmap::Beatmap(Game* g, std::filesystem::path b) : game(g), beatmap(b) {
     loadSettings();
     loadSong();
+    loadBgPicture();
     loadNotes();
-    delay = SDL_GetTicks();
+    
+    accDelay = SDL_GetTicks();
     game->setState(BEATMAP_PLAYING);
 }
 
 void Beatmap::loadSettings() {
     static Settings* settings = game->getSettings();
     speed = std::stoi(settings->get("speed"));
+    bgOpacity = static_cast<Uint8>(255 * (std::stoi(settings->get("bg_opacity")) / 100.0f));
 }
 
 void Beatmap::loadSong() {
@@ -30,6 +33,24 @@ void Beatmap::loadSong() {
             game->delay(game->getData()->offsets->beatmapStart, [this]() {
                 H2DE_PlaySound(engine, 1, song, 0);
             });
+            break;
+        }
+    }
+}
+
+void Beatmap::loadBgPicture() {
+    static H2DE_Engine* engine = game->getEngine();
+
+    if (std::filesystem::exists(beatmap)) {
+        std::ifstream difficultyFile(beatmap);
+        std::string line;
+        while (std::getline(difficultyFile, line)) {
+            size_t index = line.find(".jpg");
+            if (index == std::string::npos) continue;
+
+            bgPicture = line.substr(5, index - 5 + 4);
+            std::filesystem::path dir = beatmap.parent_path();
+            H2DE_LoadAsset(engine, dir / bgPicture);
             break;
         }
     }
@@ -78,9 +99,15 @@ void Beatmap::loadNotes() {
 // CLEANUP
 Beatmap::~Beatmap() {
     for (Note* note : notes) delete note;
+
     notes.clear();
     pressedNotes.clear();
-    // remove song from engine
+    slidersHolding.clear();
+    sliderJudgments.clear();
+    judgments.clear();
+
+    H2DE_RemoveAsset(game->getEngine(), song);
+    H2DE_RemoveAsset(game->getEngine(), bgPicture);
 }
 
 // EVENTS
@@ -95,14 +122,30 @@ void Beatmap::inputUp(int key) {
     keyReleasedThisFrame.push_back(key);
 }
 
+void Beatmap::addJudgment(Judgment judgment) {
+    static std::unordered_map<Judgment, std::string> strJudgments = {
+        { MARVELLOUS, "300+" },
+        { PERFECT, "300" },
+        { GREAT, "200" },
+        { GOOD, "100" },
+        { OK, "50" },
+        { MISS, "MISS" },
+    };
+    judgments[judgment]++;
+    thisFrameJudgment = new Judgment(judgment);
+    currentJudgment = new Judgment(judgment);
+}
+
 // UPDATE
 void Beatmap::update() {
-    static int FPS = game->getFPS();
     static int offset = game->getData()->offsets->beatmapStart;
+    static int maxJudgmentTiming = game->getData()->offsets->maxJudgmentTiming;
+
+    H2DE_TickTimelineManager(tm);
 
     // 1 => Checking for start delay
-    int currentTime = SDL_GetTicks() - delay;
-    if (currentTime < offset) return;
+    int currentTime = SDL_GetTicks() - accDelay - offset;
+    if (currentTime < 0) return;
 
     // 2 => Checking for key pressed this frame
     for (const int key : keyPressedThisFrame) {
@@ -111,18 +154,216 @@ void Beatmap::update() {
 
         int delay = std::abs(closestNote->start - static_cast<int>(currentTime));
         Judgment judgment = getJudgment(delay);
+        addJudgment(judgment);
         pressedNotes.insert(closestNote);
+
+        if (closestNote->end != 0) {
+            slidersHolding.push_back(closestNote);
+            sliderJudgments[closestNote] = judgment;
+        }
     }
 
     // 3 => Checking for key released this frame
+    for (const int key : keyReleasedThisFrame) {
+        for (int i = 0; i < slidersHolding.size(); i++) {
+            Note* note = slidersHolding[i];
 
+            if (note->column == key + 1) {
+                int delay = note->end - static_cast<int>(currentTime);
+                Judgment judgment = (delay < -maxJudgmentTiming) ? static_cast<Judgment>(sliderJudgments[note] + 1) : getJudgment(std::abs(delay));
+                if (judgment == MISS) judgment = OK;
+                addJudgment(judgment);
+
+                slidersHolding.erase(slidersHolding.begin() + i);
+                break;
+            }
+        }
+    }
+
+    // 4 => Checking for misses
+    for (Note* note : notes) {
+        if (pressedNotes.find(note) == pressedNotes.end()) {
+            int delay = note->start - static_cast<int>(currentTime);
+            if (delay < -maxJudgmentTiming) {
+                addJudgment(MISS);
+                pressedNotes.insert(note);
+            } else if (delay > maxJudgmentTiming) break;
+        }
+    }
+
+    // 5 => Reset
     keyPressedThisFrame.clear();
     keyReleasedThisFrame.clear();
 }
 
 // RENDER
 void Beatmap::render() {
+    renderBackground();
+    renderColumn();
+    renderNotes();
+    renderJudgment();
+}
 
+void Beatmap::renderBackground() {
+    static H2DE_Engine* engine = game->getEngine();
+    int winWidth, winHeight;
+    game->getWinSize(&winWidth, &winHeight);
+
+    H2DE_GraphicObject* picture = H2DE_CreateGraphicObject();
+    picture->type = IMAGE;
+    picture->pos = { 0, 0 };
+    picture->size = { winWidth, winHeight };
+    picture->texture = bgPicture;
+    picture->rgb = { bgOpacity, bgOpacity, bgOpacity, 255 };
+    picture->index = 1;
+    H2DE_AddGraphicObject(engine, picture);
+}
+
+void Beatmap::renderColumn() {
+    static H2DE_Engine* engine = game->getEngine();
+    static Calculator* calculator = game->getCalculator();
+    static GameData* gameData = game->getData();
+
+
+
+    // 1 => Render black background
+    static H2DE_Pos absNoteBgPos = calculator->convertToPx(gameData->positions->noteBg, gameData->sizes->noteBg);
+    static H2DE_Size absNoteBgSize = calculator->convertToPx(gameData->sizes->noteBg);
+
+    H2DE_GraphicObject* noteBg = H2DE_CreateGraphicObject();
+    noteBg->type = POLYGON;
+    noteBg->pos = absNoteBgPos;
+    noteBg->points = {
+        { 0, 0 },
+        { absNoteBgSize.w, 0 },
+        { absNoteBgSize.w, absNoteBgSize.h },
+        { 0, absNoteBgSize.h },
+    };
+    noteBg->rgb = { 0, 0, 0, 255 };
+    noteBg->filled = true;
+    noteBg->index = 4;
+    H2DE_AddGraphicObject(engine, noteBg);
+
+
+
+    // 2 => Render borders
+    static H2DE_Pos absNoteBgBorderPos = calculator->convertToPx(gameData->positions->noteBgBorder, gameData->sizes->noteBgBorder);
+    static H2DE_Size absNoteBgBorderSize = calculator->convertToPx(gameData->sizes->noteBgBorder);
+
+    H2DE_GraphicObject* noteBgBorder = H2DE_CreateGraphicObject();
+    noteBgBorder->type = POLYGON;
+    noteBgBorder->pos = absNoteBgBorderPos;
+    noteBgBorder->points = {
+        { 0, 0 },
+        { absNoteBgBorderSize.w, 0 },
+        { absNoteBgBorderSize.w, absNoteBgBorderSize.h },
+        { 0, absNoteBgBorderSize.h },
+    };
+    noteBgBorder->rgb = { 127, 127, 127, 255 };
+    noteBgBorder->filled = true;
+    noteBgBorder->index = 3;
+    H2DE_AddGraphicObject(engine, noteBgBorder);
+
+
+
+    // 3 => Render 4 bottom keys
+    static BeatmapPos keyPos = gameData->positions->key;
+    static BeatmapSize keySize = gameData->sizes->key;
+    static H2DE_Size absKeySize = calculator->convertToPx(keySize);
+
+    for (int i = 0; i < 4; i++) {
+        H2DE_Pos absPos = calculator->convertToPx({ i * 1.0f, keyPos.y }, keySize);
+        std::string keyID = (i == 0 || i == 3) ? "1" : "2";
+        std::string keyState = (keysDown[i]) ? "D" : "";
+
+        H2DE_GraphicObject* key = H2DE_CreateGraphicObject();
+        key->type = IMAGE;
+        key->pos = absPos;
+        key->size = absKeySize;
+        key->texture = "mania-key" + keyID + keyState + ".png";
+        key->index = 10;
+        H2DE_AddGraphicObject(engine, key);
+    }
+
+
+
+    // 4 => Render key top bar
+    static H2DE_Pos absKeyBarPos = calculator->convertToPx(gameData->positions->keyBar, gameData->sizes->keyBar);
+    static H2DE_Size absKeyBarSize = calculator->convertToPx(gameData->sizes->keyBar);
+
+    H2DE_GraphicObject* keyBar = H2DE_CreateGraphicObject();
+    keyBar->type = IMAGE;
+    keyBar->pos = absKeyBarPos;
+    keyBar->size = absKeyBarSize;
+    keyBar->texture = "mania-stage-hint.png";
+    keyBar->index = 11;
+    H2DE_AddGraphicObject(engine, keyBar);
+
+
+
+    // 5 => Render health bar
+    static H2DE_Pos absHealthBarPos = calculator->convertToPx(gameData->positions->healthBar, gameData->sizes->healthBar);
+    static H2DE_Size absHealthBarSize = calculator->convertToPx(gameData->sizes->healthBar);
+
+    H2DE_GraphicObject* healthBar = H2DE_CreateGraphicObject();
+    healthBar->type = IMAGE;
+    healthBar->pos = absHealthBarPos;
+    healthBar->size = absHealthBarSize;
+    healthBar->texture = "scorebar-bg@2x.png";
+    healthBar->rotation = 270;
+    healthBar->index = 15;
+    H2DE_AddGraphicObject(engine, healthBar);
+}
+
+void Beatmap::renderNotes() {
+    static H2DE_Engine* engine = game->getEngine();
+    static Calculator* calculator = game->getCalculator();
+
+
+    // BeatmapPos pos = { 0.0f, 0.0f };
+    // BeatmapSize size = { 1.0f, 0.32f };
+
+    // H2DE_GraphicObject* note = H2DE_CreateGraphicObject();
+    // note->type = IMAGE;
+    // note->pos = calculator->convertToPx(pos, size);
+    // note->size = calculator->convertToPx(size);
+    // note->texture = "mania-note1.png";
+    // note->index = 5;
+    // H2DE_AddGraphicObject(engine, note);
+}
+
+void Beatmap::renderJudgment() {
+    static H2DE_Engine* engine = game->getEngine();
+    static Calculator* calculator = game->getCalculator();
+    static GameData* gameData = game->getData();
+
+    if (!currentJudgment) return;
+
+    BeatmapSize judgmentSize = gameData->sizes->judgments[*currentJudgment];
+    H2DE_Pos absJudgmentPos = calculator->convertToPx(gameData->positions->judgment, judgmentSize);
+    H2DE_Size absJudgmentSize = calculator->convertToPx(judgmentSize);
+
+    H2DE_GraphicObject* judgment = H2DE_CreateGraphicObject();
+    judgment->type = IMAGE;
+    judgment->pos = absJudgmentPos;
+    judgment->size = absJudgmentSize;
+    judgment->texture = gameData->others->stringifiedJudgment[*currentJudgment] + ".png";
+    judgment->scaleOrigin = { static_cast<int>(absJudgmentSize.w / 2), static_cast<int>(absJudgmentSize.h / 2) };
+    judgment->scale = { currentJudgmentScale, currentJudgmentScale };
+    judgment->index = 20;
+    H2DE_AddGraphicObject(engine, judgment);
+
+    if (!thisFrameJudgment) return;
+
+    H2DE_AddTimelineToManager(tm, H2DE_CreateTimeline(engine, 50, EASE_OUT, [this](float blend) {
+        currentJudgmentScale = 1.0f + blend * (1.5f - 1.0f);
+    }, [this]() {
+        delete currentJudgment;
+        currentJudgment = nullptr;
+    }, 0));
+
+    delete thisFrameJudgment;
+    thisFrameJudgment = nullptr;
 }
 
 // GETTER