+++ src/functions.cpp
@@ -1,1 +1,5 @@
 #include "functions.h"
+
+float lerp(float min, float max, float blend) {
+    return min + blend * (max - min);
+}