+++ CHANGES.md
@@ -1,3 +1,17 @@
+## HOM-1.0.2
+#### ADDED
+- sliders
+- misses
+- beatmap background picture
+- beatmap background picture opacity
+- beatmap background picture opacity setting
+- default skin textures
+- judgments visuals
+#### FIXED
+- can't hit notes in the first 3 seconds
+#### LIB
+- updated H2DE from version 1.3.10 to 1.3.12
+
 ## HOM-1.0.1
 #### ADDED
 - opening beatmap from menu
@@ -5,7 +19,7 @@
 - key inputs
 - input detection to closest note with judgments
 #### LIB
-- updated H2DE from version 1.3.10 to 1.3.11
+- updated H2DE from version 1.3.9 to 1.3.10
 
 ## HOM-1.0.0
 #### ADDED